Significant update to the publisher_mapping.csv mapping file.
This should be a more sustainable mapping.
The problem is that there is no machine mapping available for the
mappings required by the transparency tests - e.g. Signatory
so a manually maintained list must be kept. This is a pain!

With this set up we can get new publishers onto the list easily,
and if we don't designate a 'type' then it will be assigned as other.

The other dignificant part of the mapping is the Shortname section that
has no machine readable equivalent elsewhere.